fix appcleaner to remove only user-application containers
in docker <= 1.9 the 'ancestor' filter didn't exist, so docker was
ignoring it and removing all exited containers wether they were from
users or from backend services. Also, the open365-app-docker image has
not existed for quite some time, so in docker >= 1.9 what happened was
that no image was being removed.

We have fixed everything by setting a label only to user-containers and
just removing the exited containers that have that label.